Separate click UI from CLI logic, in preparation for testing
Wanted to test the CLI in isolation (instead of the current
`test_main.py`. Needed to separate out the `click` UI logic and make the
AnagramFinder path configurable in preparation for a mocked REST server test.

NOTE: this has increased the number of validations that need to be incorporated
into the non-`click` UI logic.