Created a tornado `main.py` to compare against the Flask version
Found out that Flask is synchronous and haven't found a good enough answer yet
to make it async (searching, seems to suggest using `asyncio`/`aiohttp`, or
`tornado` or `flask`/`tornado`).

TEMP: Duplicated `main.py` to `main_tornado.py` so that I could get a side by
side comparison of the differences in the two libraries.